content: Reorder actions and group per tool
Groups actions related to each of the tools that get installed by this
stage: README, Kolibri for Windows, USB Launcher and Autorun data. While
at it, this also tries to highlight a dependency that the USB Laucher
has on kolibri-windows.zip.

This commit changes the order in which actions occur, but should not
bring any functional change.

https://phabricator.endlessm.com/T33670