Add configuration support
Wrap the main build script with a python program that can parse
ini-style config files, and add support for the settings we'll need to
build different variations of the Endless Key. This also renames some
configuration variables to match this new configuration scheme.

Most of this logic was copied from eos-image-builder.

https://phabricator.endlessm.com/T33670