kolibri: Parametrize the prefix for the PROJECT setting
The PROJECT setting in options.ini can be used as a filter for Learning
Equality's metrics. This commit adds a command line parameter to allow
defining the prefix used in the value for that setting. The build
configuration codename is then appended to this prefix to form the full
string for the PROJECT setting.

It also changes the Jenkins job to make use of this new parameter and
set the "endless-key" prefix, which we have been using to identify
official Endless Key builds.

https://phabricator.endlessm.com/T33787